+++ src/mail-scheduler/tests/run-tests.sh
@@ -6,9 +6,9 @@ cd $SCRIPT_PATH/../../..
 
 if ! [[ "$(docker container inspect -f '{{.State.Running}}' orion_db)" = "true" ]]; then
     docker network create joystream_default || true
-    docker-compose up -d orion_db
-    docker-compose up -d orion_processor # We need the processor to run the migrations
-    until docker-compose logs orion_db | grep "database system is ready to accept connections" >/dev/null; do
+    docker compose up -d orion_db
+    docker compose up -d orion_processor # We need the processor to run the migrations
+    until docker compose logs orion_db | grep "database system is ready to accept connections" >/dev/null; do
         sleep 1
     done
 fi
@@ -16,7 +16,7 @@ fi
 sleep 10
 
 cleanup() {
-    docker-compose down -v
+    docker compose down -v
 }
 
 # Run the tests