+++ src/tests/integration/run.sh
@@ -6,9 +6,9 @@ cd $SCRIPT_PATH/../../..
 
 if ! [[ "$(docker container inspect -f '{{.State.Running}}' orion_db)" = "true" ]]; then
     docker network create joystream_default || true
-    docker-compose up -d orion_db &
+    docker compose up -d orion_db &
     2>/dev/null
-    until docker-compose logs orion_db | grep "database system is ready to accept connections" >/dev/null; do
+    until docker compose logs orion_db | grep "database system is ready to accept connections" >/dev/null; do
         echo "Waiting for the db to be ready..."
         sleep 1
     done
@@ -16,7 +16,7 @@ if ! [[ "$(docker container inspect -f '{{.State.Running}}' orion_db)" = "true" 
 fi
 
 cleanup() {
-    docker-compose down -v
+    docker compose down -v
 }
 
 # Run the tests